Fix issues in install_dev_extension.sh

+++ dev/install_dev_extension.sh
@@ -6,12 +6,12 @@ EXTENSION_ID=com.github.kbialek.ulauncher-bitwarden
 ULAUNCHER_EXT_DIR=~/.local/share/ulauncher/extensions/
 
 # Remove whatever version of this extension is installed
-rm -rf ${ULAUNCHER_EXT_DIR}/${EXTENSION_ID}
+rm -rf ${ULAUNCHER_EXT_DIR:?}/${EXTENSION_ID}
 
 # Symlink current dir if extension is here
 if [ -e "manifest.json" ]
 then
-	ln -s $(pwd) ${ULAUNCHER_EXT_DIR}/${EXTENSION_ID}
+	ln -s "$(pwd)" ${ULAUNCHER_EXT_DIR}/${EXTENSION_ID}
 	echo "Done"
 else
 	echo "Please run this script from the root directory of the extension"
